Fix filtering scripts and update 2018 json files
